Implemented permissions, havent tested yet

+++ T2/udpserver.c
@@ -27,7 +27,7 @@ struct stat info={0};
 
 int numUsers = 6;
 const char * existingUsers[] = {"user1", "user2", "user3", "user4", "user5", "user6"};
-const int existingPasswords[] = {210723987853, 210723987854, 210723987854, 210723987855, 210723987856, 210723987854};
+const int existingPasswords[] = {210723987853, 210723987854, 210723987855, 210723987856, 210723987857, 210723987858};
 const char * validationTokens[6];
 int currentUser;
 
@@ -209,11 +209,77 @@ char * readFile(char * path, int nrbytes, int offset)
   return readMachinetoHuman(payload,tam);
 }
 
+void createFilePermission(char * path, int ownerPerm, int otherPerm){
+	int fd;
+	char * base, * basec, * dirc, * dname;
+  	char hidden[80], hiddenPath[80], perms[10];
+
+  	basec = strdup(path);
+  	base = basename(basec);
+  	dirc = strdup(path);
+  	dname = dirname(dirc);
+
+  	sprintf(hidden, ".%s", base);
+	sprintf(hiddenPath, "%s/%s", dname, hidden);
+	sprintf(perms, "%d %d %d", currentUser, ownerPerm, otherPerm);
+
+	fd = open(hiddenPath, O_CREAT|O_RDWR, PERALL);
+	write(fd, perms , sizeof(perms));
+	close(fd);
+}
+
+int checkFilePermission(char * path){
+	int fd;
+	char * base, * basec, * dirc, * dname;
+  	char hidden[80], hiddenPath[80], buffer[10];
+
+  	basec = strdup(path);
+  	base = basename(basec);
+  	dirc = strdup(path);
+  	dname = dirname(dirc);
+
+  	sprintf(hidden, ".%s", base);
+	sprintf(hiddenPath, "%s/%s", dname, hidden);
+	fd = open(hiddenPath, O_CREAT|O_RDWR, PERALL);
+
+	read(fd, buffer, 1);
+	if(atoi(buffer) != currentUser){
+		lseek(fd, 4, SEEK_SET);
+	}
+	else{
+		lseek(fd, 2, SEEK_SET);
+	}
+	read(fd, buffer, 1);
+	printf("%s\n", buffer);
+	close(fd);
+	return atoi(buffer);
+}
+
+int checkDirPermission(char * path, char * dirname){
+	char hidden[80], hiddenPath[80], buffer[10];
+	int fd;
+
+	sprintf(hidden, ".%s", dirname);
+  	sprintf(hiddenPath, "%s/%s", path, hidden);
+  	fd = open(hiddenPath, O_CREAT|O_RDWR, PERALL);
+  	read(fd, buffer, 1);
+	if(atoi(buffer) != currentUser){
+		lseek(fd, 4, SEEK_SET);
+	}
+	else{
+		lseek(fd, 2, SEEK_SET);
+	}
+	read(fd, buffer, 1);
+	printf("%s\n", buffer);
+	close(fd);
+	return atoi(buffer);
+}
+
 /* Escreve uma quantidade de bytes de um arquivo a partir de um offset */
 char * writeFile(char * path, char * payload, int nrbytes, int offset, int ownerPerm, int otherPerm)
 { 
   int fd;
-
+  
   printf("\n\nOperacao: escrita da string >>%s<< no arquivo >>%s<< a partir do offset %d.\n", payload, getName(path), offset);
 
 
@@ -221,6 +287,15 @@ char * writeFile(char * path, char * payload, int nrbytes, int offset, int owner
   {
     unlink(path);
     return writeMachinetoHuman("Você chamou a função que escreve em arquivos com nrbytes=0, seu arquivo foi apagado");
+  }
+
+  if( access( path, F_OK ) == -1 ) {
+  	createFilePermission(path, ownerPerm, otherPerm);
+  }
+  else{
+  	if(!checkFilePermission(path)){
+  		return erros("Você não tem permissão para escrever neste arquivo.");
+  	}
   }
 
   fd = open(path, O_CREAT|O_RDWR, PERALL);// cria se nao existir e seta a permissao libero geral
@@ -234,8 +309,6 @@ char * writeFile(char * path, char * payload, int nrbytes, int offset, int owner
     lseek(fd, offset, SEEK_SET);
 
   write(fd, payload, nrbytes);
-
-
   close(fd);
 
   return writeMachinetoHuman("Sua mensagem foi escrita.");
@@ -273,6 +346,8 @@ char * fileInfo(char * path)
 char * makdir(char * path, char * dirname, int ownerPerm, int otherPerm)
 {
   char mk[BUFSIZE];
+  char hidden[80], hiddenPath[80], perms[10];
+  int fd;
 
   printf("\n\nOperacao: criar um diretório de nome >>%s<< em %s.\n", dirname, path);
 
@@ -285,6 +360,15 @@ char * makdir(char * path, char * dirname, int ownerPerm, int otherPerm)
     sprintf(mk,"Erro na criacao do diretorio >>%s<<.", dirname);
     return erros(mk);
   }
+
+  // create permission
+  sprintf(hidden, ".%s", dirname);
+  sprintf(hiddenPath, "%s/%s", path, hidden);
+  sprintf(perms, "%d %d %d", currentUser, ownerPerm, otherPerm);
+
+  fd = open(hiddenPath, O_CREAT|O_RDWR, PERALL);
+  write(fd, perms, sizeof(perms));
+  close(fd);
 
   sprintf(mk,"Diretório criado, novo path: %s.",mk);
 
@@ -300,6 +384,10 @@ char * rm(char * path, char * dirname)
   strcat(mk, dirname);
 
   printf("\n\nOperação: remover o diretório de nome >>%s<< em %s.\n", dirname, path);
+
+  if(!checkDirPermission(path, dirname)){
+  		return erros("Você não tem permissão para remover este diretório.");
+  }
   
   if( (rmdir(mk)) == -1)
   {
